chore: move post message service to /service
